backend/requirements.txt
fastapi==0.115.12
openai==1.76.0
pillow==11.2.1
pydantic==2.11.4
python-multipart==0.0.20
uvicorn[standard]==0.34.2
